# perf: reduce allocations and boxing