--- a/YoloDotNet/Models/IDetection.cs
+++ b/YoloDotNet/Models/IDetection.cs
@@ -1,9 +1,0 @@
-﻿namespace YoloDotNet.Models
-{
-    public interface IDetection
-    {
-        LabelModel Label { get; init; }
-        double Confidence { get; init; }
-        SKRectI BoundingBox { get; init; }
-    }
-}