--- a/YoloDotNet/Models/KeyPoint.cs
+++ b/YoloDotNet/Models/KeyPoint.cs
@@ -1,4 +1,17 @@
 ﻿namespace YoloDotNet.Models
 {
-    public record KeyPoint(int X, int Y, double Confidence);
+    public struct KeyPoint
+    {
+        public KeyPoint(int x, int y, double confidence)
+        {
+            this.X = x;
+            this.Y = y;
+            this.Confidence = confidence;
+        }
+
+        public int X { get; }
+        public int Y { get; }
+        public double Confidence { get; }
+    }
+
 }