--- a/YoloDotNet/Models/OBBDetection.cs
+++ b/YoloDotNet/Models/OBBDetection.cs
@@ -3,22 +3,15 @@
     /// <summary>
     /// Represents the result of object detection, including label information, confidence score, and bounding box.
     /// </summary>
-    public class OBBDetection : IDetection
+    public struct OBBDetection
     {
-        /// <summary>
-        /// Label information associated with the detected object.
-        /// </summary>
-        public LabelModel Label { get; init; } = new();
+        public OBBDetection(Detection detection, float orientationAngle)
+        {
+            Detection = detection;
+            OrientationAngle = orientationAngle;
+        }
 
-        /// <summary>
-        /// Confidence score of the detected object.
-        /// </summary>
-        public double Confidence { get; init; }
-
-        /// <summary>
-        /// Rectangle defining the region of interest (bounding box) of the detected object.
-        /// </summary>
-        public SKRectI BoundingBox { get; init; }
+        public Detection Detection { get; }
 
         /// <summary>
         /// Orientation angle of bounding box