--- a/YoloDotNet/Models/ObjectDetection.cs
+++ b/YoloDotNet/Models/ObjectDetection.cs
@@ -3,22 +3,14 @@
     /// <summary>
     /// Represents the result of object detection, including label information, confidence score, and bounding box.
     /// </summary>
-    public class ObjectDetection : IDetection
+    public struct ObjectDetection
     {
-        /// <summary>
-        /// Label information associated with the detected object.
-        /// </summary>
-        public LabelModel Label { get; init; } = new();
+        public ObjectDetection(Detection detection)
+        {
+            Detection = detection;
+        }
 
-        /// <summary>
-        /// Confidence score of the detected object.
-        /// </summary>
-        public double Confidence { get; init; }
-
-        /// <summary>
-        /// Rectangle defining the region of interest (bounding box) of the detected object.
-        /// </summary>
-        public SKRectI BoundingBox { get; init; }
+        public Detection Detection { get; }
 
     }
 }