--- a/YoloDotNet/Modules/Interfaces/IClassificationModule.cs
+++ b/YoloDotNet/Modules/Interfaces/IClassificationModule.cs
@@ -2,7 +2,7 @@
 {
     public interface IClassificationModule : IModule
     {
-        List<Classification> ProcessImage(SKImage image, double classes, double pixelConfidence, double iou);
-        Dictionary<int, List<Classification>> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou);
+        Classification[] ProcessImage(SKImage image, double classes, double pixelConfidence, double iou);
+        Dictionary<int, Classification[]> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou);
     }
 }