--- a/YoloDotNet/Modules/Interfaces/IObjectDetectionModule.cs
+++ b/YoloDotNet/Modules/Interfaces/IObjectDetectionModule.cs
@@ -2,7 +2,7 @@
 {
     public interface IObjectDetectionModule : IModule
     {
-        List<ObjectDetection> ProcessImage(SKImage image, double confidence, double pixelConfidence,double iou);
-        Dictionary<int, List<ObjectDetection>> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou);
+        ObjectDetection[] ProcessImage(SKImage image, double confidence, double pixelConfidence,double iou);
+        Dictionary<int, ObjectDetection[]> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou);
     }
 }