--- a/YoloDotNet/Modules/Interfaces/IPoseEstimationModule.cs
+++ b/YoloDotNet/Modules/Interfaces/IPoseEstimationModule.cs
@@ -2,7 +2,7 @@
 {
     internal interface IPoseEstimationModule : IModule
     {
-        List<PoseEstimation> ProcessImage(SKImage image, double confidence, double pixelConfidence, double iou);
-        Dictionary<int, List<PoseEstimation>> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou);
+        PoseEstimation[] ProcessImage(SKImage image, double confidence, double pixelConfidence, double iou);
+        Dictionary<int, PoseEstimation[]> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou);
     }
 }