--- a/YoloDotNet/Modules/Interfaces/ISegmentationModule.cs
+++ b/YoloDotNet/Modules/Interfaces/ISegmentationModule.cs
@@ -2,7 +2,7 @@
 {
     public interface ISegmentationModule : IModule
     {
-        List<Segmentation> ProcessImage(SKImage image, double confidence, double pixelConfidence, double iou);
-        Dictionary<int, List<Segmentation>> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou);
+        Segmentation[] ProcessImage(SKImage image, double confidence, double pixelConfidence, double iou);
+        Dictionary<int, Segmentation[]> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou);
     }
 }