--- a/YoloDotNet/Modules/V10/ObjectDetectionModuleV10.cs
+++ b/YoloDotNet/Modules/V10/ObjectDetectionModuleV10.cs
@@ -18,7 +18,7 @@ namespace YoloDotNet.Modules.V10
             SubscribeToVideoEvents();
         }
 
-        public List<ObjectDetection> ProcessImage(SKImage image, double confidence, double pixelConfidence, double iou)
+        public ObjectDetection[] ProcessImage(SKImage image, double confidence, double pixelConfidence, double iou)
         {
             using var ortValues = _yoloCore!.Run(image);
             using var ort = ortValues[0];
@@ -28,7 +28,7 @@ namespace YoloDotNet.Modules.V10
             return [.. results];
         }
 
-        public Dictionary<int, List<ObjectDetection>> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou)
+        public Dictionary<int, ObjectDetection[]> ProcessVideo(VideoOptions options, double confidence, double pixelConfidence, double iou)
             => _yoloCore.RunVideo(options, confidence, pixelConfidence, iou, ProcessImage);
 
         #region Helper methods