Reduce allocations in the following ways

1. Structs to replace classes
2. Use Array<T> not List<T>
3. Prevent boxing to an interface

These are the results from my computer.

Starting Values:

| Method                    | Mean     | Error    | StdDev   | Allocated |
|-------------------------- |---------:|---------:|---------:|----------:|
| ObjectDetectionYolov8Cpu  | 66.22 ms | 1.754 ms | 5.173 ms |  34.48 KB |
| ObjectDetectionYolov8Gpu  | 11.64 ms | 0.174 ms | 0.145 ms |  34.48 KB |
| ObjectDetectionYolov9Cpu  | 67.42 ms | 1.263 ms | 2.403 ms |   29.4 KB |
| ObjectDetectionYolov9Gpu  | 17.66 ms | 0.044 ms | 0.039 ms |  29.33 KB |
| ObjectDetectionYolov10Cpu | 65.90 ms | 1.303 ms | 3.197 ms |  24.26 KB |
| ObjectDetectionYolov10Gpu | 10.35 ms | 0.041 ms | 0.034 ms |   24.2 KB |
| ObjectDetectionYolov11Cpu | 71.02 ms | 0.643 ms | 0.570 ms |  32.65 KB |
| ObjectDetectionYolov11Gpu | 11.59 ms | 0.035 ms | 0.031 ms |  32.58 KB |
| ObjectDetectionYolov12Cpu | 85.72 ms | 1.706 ms | 3.598 ms |  31.05 KB |
| ObjectDetectionYolov12Gpu | 13.82 ms | 0.075 ms | 0.066 ms |  30.96 KB |

Ending Values: 

| Method                    | Mean     | Error    | StdDev   | Allocated |
|-------------------------- |---------:|---------:|---------:|----------:|
| ObjectDetectionYolov8Cpu  | 62.00 ms | 1.223 ms | 2.110 ms |  27.99 KB |
| ObjectDetectionYolov8Gpu  | 11.55 ms | 0.019 ms | 0.016 ms |  27.93 KB |
| ObjectDetectionYolov9Cpu  | 65.64 ms | 1.284 ms | 1.319 ms |   23.9 KB |
| ObjectDetectionYolov9Gpu  | 17.63 ms | 0.084 ms | 0.079 ms |  23.83 KB |
| ObjectDetectionYolov10Cpu | 54.61 ms | 0.884 ms | 0.784 ms |  22.21 KB |
| ObjectDetectionYolov10Gpu | 10.39 ms | 0.100 ms | 0.094 ms |  22.14 KB |
| ObjectDetectionYolov11Cpu | 61.14 ms | 1.212 ms | 2.123 ms |  26.52 KB |
| ObjectDetectionYolov11Gpu | 11.56 ms | 0.064 ms | 0.056 ms |  26.45 KB |
| ObjectDetectionYolov12Cpu | 73.53 ms | 1.416 ms | 1.325 ms |  25.17 KB |
| ObjectDetectionYolov12Gpu | 13.68 ms | 0.057 ms | 0.053 ms |  25.08 KB |

The benchmarks only run the method once, so the GC impact is not showing.
This is something that can be added in future.

All tests passing.
Demo's still work as intended.