@louislewis2 2025-07-16
@NickSwardh Good morning.

I took a look through the code.
My main changes will not work with the updates to the model, specifically the implementation of the tracking.

Previously the models were assigned to once, that along with the fact that the overall memory size of the classes along with its properties where small enough to warrant moving over to a struct.

The issue I was resolving, could be seen where the inference service is continuously processes.
A benchmark that runs inference around 1024 times and greater shows the GC impact I was resolving.
This greatly effects systems that are long running i.e never ending.

Below is a snippet of the benchmark results.

Starting Point:

| Method    | FrameCount | Mean         | Error      | StdDev     | Gen0      | Allocated    |
|---------- |----------- |-------------:|-----------:|-----------:|----------:|-------------:|
| Inference | 1          |     44.88 ms |   0.893 ms |   1.416 ms |         - |    572.24 KB |
| Inference | 2          |     92.27 ms |   1.841 ms |   4.411 ms |         - |   1144.48 KB |
| Inference | 4          |    178.99 ms |   3.542 ms |   6.014 ms |         - |   2289.22 KB |
| Inference | 8          |    356.79 ms |   6.996 ms |  11.880 ms |         - |   4577.94 KB |
| Inference | 16         |    708.59 ms |  14.045 ms |  25.325 ms |         - |   9155.88 KB |
| Inference | 32         |  1,429.98 ms |  28.248 ms |  38.667 ms |         - |  18311.75 KB |
| Inference | 64         |  2,897.98 ms |  39.696 ms |  35.189 ms |         - |   36623.5 KB |
| Inference | 128        |  6,093.04 ms | 121.557 ms | 174.334 ms |         - |     73247 KB |
| Inference | 256        | 11,622.18 ms |  58.579 ms |  54.794 ms |         - | 146494.09 KB |
| Inference | 512        | 24,597.57 ms | 482.520 ms | 870.083 ms |         - | 292989.09 KB |
| Inference | 1024       | 46,297.59 ms | 140.258 ms | 117.122 ms | 1000.0000 | 585978.48 KB |

Ending Point:

| Method    | FrameCount | Mean         | Error      | StdDev     | Median       | Allocated    |
|---------- |----------- |-------------:|-----------:|-----------:|-------------:|-------------:|
| Inference | 1          |     42.21 ms |   0.835 ms |   1.143 ms |     42.65 ms |    204.55 KB |
| Inference | 2          |     84.08 ms |   1.663 ms |   3.282 ms |     85.61 ms |    409.11 KB |
| Inference | 4          |    166.08 ms |   3.310 ms |   6.135 ms |    167.73 ms |    818.22 KB |
| Inference | 8          |    328.50 ms |   6.475 ms |  11.509 ms |    334.36 ms |   1636.44 KB |
| Inference | 16         |    659.77 ms |  12.982 ms |  23.076 ms |    670.83 ms |   3273.89 KB |
| Inference | 32         |  1,325.71 ms |  25.857 ms |  36.248 ms |  1,339.85 ms |   6554.19 KB |
| Inference | 64         |  2,673.58 ms |  49.431 ms |  48.547 ms |  2,684.89 ms |   13091.5 KB |
| Inference | 128        |  5,489.17 ms |  92.764 ms |  86.772 ms |  5,494.93 ms |     26183 KB |
| Inference | 256        | 10,905.75 ms | 152.240 ms | 134.957 ms | 10,974.53 ms |  52366.27 KB |
| Inference | 512        | 21,750.79 ms | 312.611 ms | 292.416 ms | 21,577.53 ms | 104737.49 KB |
| Inference | 1024       | 43,913.61 ms | 545.309 ms | 483.402 ms | 44,123.49 ms | 209465.38 KB |

Note the Gen0 value.